+++ Lib/fontbakery/profiles/googlefonts.py
@@ -1854,7 +1854,7 @@ def com_google_fonts_check_metadata_familyname(family_metadata):
     conditions = ['family_metadata']
 )
 def com_google_fonts_check_metadata_has_regular(family_metadata):
-    """METADATA.pb: According Google Fonts standards,
+    """METADATA.pb: According to Google Fonts standards,
        families should have a Regular style.
     """
     from .googlefonts_conditions import has_regular_style
@@ -1866,7 +1866,10 @@ def com_google_fonts_check_metadata_has_regular(family_metadata):
               Message("lacks-regular",
                       "This family lacks a Regular"
                       " (style: normal and weight: 400)"
-                      " as required by Google Fonts standards.")
+                      " as required by Google Fonts standards."
+                      " If family consists of a single-weight non-Regular style only,"
+                      " consider the Google Fonts specs for this case:"
+                      " https://github.com/googlefonts/gf-docs/tree/main/Spec#single-weight-families")
 
 
 @check(